examples/python/requirements.txt
--index-url=https://blpapi.bloomberg.com/repository/releases/python/simple/
blpapi
python-dotenv
